pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "healco-lite"
version = "0.1.0"
description = "Simplified HealCo Telegram bot for counting calories, menus, workouts, and motivation"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Rafael Sayadi"}]
dependencies = [
    "python-telegram-bot[webhooks]==21.4",
    "openai==1.40.2",
    "aiohttp==3.9.5",
    "python-dotenv==1.0.1",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "requests-oauthlib>=1.3.1",
    "beautifulsoup4>=4.12.2",
    "pydantic==2.7.4",
    "pydantic-core==2.18.4",
    "openfoodfacts==2.9.0",
    "psycopg2-binary>=2.9",
]

[project.scripts]
healco-lite = "main:main"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
]

[tool.setuptools]
py-modules = ["main", "search"]
